테스트 중복·스타일 정리: 커서 헬퍼 승격, 불리언 비교 통일
- 6개 파일에 바이트 단위 동일하게 복붙된 create_cursor_at_position을
  tests/utils로 승격하고 사용처를 import로 교체
- == True/False 비교 11건을 직접 진리값 검사로 통일 (E712)
