+++ tests/integration/test_lifecycle_profiler.py
@@ -10,7 +10,7 @@ from server import app
 from data.event import ServerEvent, ClientEvent
 from data.board import Point, Section, SectionFlag
 from data.board.cursorboard import Color
-from tests.utils import PytestTCM, assert_wait_event, build_tiles
+from tests.utils import PytestTCM, assert_wait_event, build_tiles, create_cursor_at_position
 from config import BoardConfig
 from core.lifecycle import LifecycleProfiler
 
@@ -34,27 +34,6 @@ async def opened_board_map(db):
         await create_section(db, section)
 
 
-def create_cursor_at_position(pos: Point):
-    """Cursor를 특정 위치에 생성하는 헬퍼"""
-    from data.cursor import Cursor
-    origin_create = Cursor.create
-
-    def create_cursor_effect(
-        id: str,
-        width: int = 0,
-        height: int = 0,
-        color: Color = Color.RED,
-        **_kwargs,
-    ):
-        return origin_create(
-            id,
-            width=width,
-            height=height,
-            position=pos,
-            color=color,
-        )
-
-    return create_cursor_effect
 
 
 @patch.object(BoardConfig, "LENGTH", new=5)
