+++ tests/utils/__init__.py
@@ -6,6 +6,7 @@ from .internal.wait_call import (
     assert_wait_event,
 )
 from .internal.builder import build_tiles
+from .internal.cursor_helpers import create_cursor_at_position
 
 
 class TestCase:
